+++ packages/taiko-client-rs/crates/protocol/src/preconfirmation/lookahead/resolver/mod.rs
@@ -1,0 +1,10 @@
+//! Lookahead resolver submodules split for readability.
+
+mod core;
+mod epoch;
+mod timeline;
+mod types;
+
+pub use core::LookaheadResolver;
+pub(crate) use epoch::MAX_LOOKBACK_EPOCHS;
+pub use types::LookaheadBroadcast;
